finish all feedback

+++ Hospital.Entities/Common/PatientListDto.cs
@@ -11,7 +11,7 @@ namespace Hospital.Entities.Common
         public int Id { get; set; }
         public string InternalNumber { get; set; }
         public string Name { get; set; }
-        public int? Age { get; set; }
+        public double? Age { get; set; }
         public string Governorate { get; set; }
         public string Gender { get; set; }
         public string Phone1 { get; set; }
